[Bug/json_float_error/BakJH]

+++ slice/InfoWriter.h
@@ -17,10 +17,10 @@ namespace Hix
 			std::filesystem::path _outfoldername;
 			std::filesystem::path _infoJsonName;
 			size_t _resX, _resY;
-			float _layerHeight;
+			double _layerHeight;
 
 		public:
-			InfoWriter(std::filesystem::path outFolder, size_t resX, size_t resY, float layerHeight);
+			InfoWriter(std::filesystem::path outFolder, size_t resX, size_t resY, double layerHeight);
 			void createInfoFile();
 			void writeVittroOptions(int max_slices, const Hix::Engine3D::Bounds3D& bound);
 			void writeBasicInfo(int sliceCnt, const std::optional<rapidjson::Value>& optionalVal);
